LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/items/column-automatic-minimum-for-auto-expected.txt

XXXXXX
XX
XX XX XX
XXX XX X
X X X X X X
XX
XXXX XXXX
XX XXXXX X
XX
XXXXXX
XXXX XXXX
XX
XX XXXXXXXX
XXXX XXXX
XX XX
XXXXXX X XXX
XX XX XX
X X
XXXX
X XXX XX
XXX XXXX
XX XX XX
XXXXX X XXXXX
X XXX XX
XXXXX XXXXX
XX XX
XX XX XX XX
XXX
X XX XXX
X XX X
XXX X
XXX X
XXX X
XX XXX
XX XXX XX
XX XXX XX
XXXXXX
XXXXXX
XXXXXX
XX XXX X
XX XXX X
XX XXX X
X XXXX X
X XXXX X
X XXXX X

FAIL Check that min-width is honored when sizing auto columns. assert_equals: expected "40px" but got "auto"
FAIL Check that min-width is honored when sizing auto columns and spanning grid items. assert_equals: expected "35px 15px" but got "auto auto"
FAIL Check the interactions between width and min-width and auto tracks. assert_equals: expected "60px" but got "auto"
FAIL Check that borders and paddings are considering when computing min sizes. assert_equals: expected "50px" but got "auto"
